LayoutTests/fast/selectors/nth-child-basics-expected.txt

Test basic uses cases of :nth-child().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ":nth-child(8)"
PASS document.querySelectorAll('#test-root :nth-child(8)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(8)')[0].id is "testcase8"
PASS styledElements is ["testcase8"]

Testing ":nth-child(n+6)"
PASS document.querySelectorAll('#test-root :nth-child(n+6)').length is 6
PASS document.querySelectorAll('#test-root :nth-child(n+6)')[0].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(n+6)')[1].id is "testcase7"
PASS document.querySelectorAll('#test-root :nth-child(n+6)')[2].id is "testcase8"
PASS document.querySelectorAll('#test-root :nth-child(n+6)')[3].id is "testcase9"
PASS document.querySelectorAll('#test-root :nth-child(n+6)')[4].id is "testcase10"
PASS document.querySelectorAll('#test-root :nth-child(n+6)')[5].id is "testcase11"
PASS styledElements is ["testcase6","testcase7","testcase8","testcase9","testcase10","testcase11"]

Testing ":nth-child(-n+9)"
PASS document.querySelectorAll('#test-root :nth-child(-n+9)').length is 9
PASS document.querySelectorAll('#test-root :nth-child(-n+9)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(-n+9)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(-n+9)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(-n+9)')[3].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(-n+9)')[4].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(-n+9)')[5].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(-n+9)')[6].id is "testcase7"
PASS document.querySelectorAll('#test-root :nth-child(-n+9)')[7].id is "testcase8"
PASS document.querySelectorAll('#test-root :nth-child(-n+9)')[8].id is "testcase9"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4","testcase5","testcase6","testcase7","testcase8","testcase9"]

Testing ":nth-child(n+4):nth-child(-n+8)"
PASS document.querySelectorAll('#test-root :nth-child(n+4):nth-child(-n+8)').length is 5
PASS document.querySelectorAll('#test-root :nth-child(n+4):nth-child(-n+8)')[0].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(n+4):nth-child(-n+8)')[1].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(n+4):nth-child(-n+8)')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(n+4):nth-child(-n+8)')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :nth-child(n+4):nth-child(-n+8)')[4].id is "testcase8"
PASS styledElements is ["testcase4","testcase5","testcase6","testcase7","testcase8"]

Testing ":nth-child(-n+8):nth-child(n+4)"
PASS document.querySelectorAll('#test-root :nth-child(-n+8):nth-child(n+4)').length is 5
PASS document.querySelectorAll('#test-root :nth-child(-n+8):nth-child(n+4)')[0].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(-n+8):nth-child(n+4)')[1].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(-n+8):nth-child(n+4)')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(-n+8):nth-child(n+4)')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :nth-child(-n+8):nth-child(n+4)')[4].id is "testcase8"
PASS styledElements is ["testcase4","testcase5","testcase6","testcase7","testcase8"]

Testing ":nth-child(n+2):nth-child(odd):nth-child(-n+9)"
PASS document.querySelectorAll('#test-root :nth-child(n+2):nth-child(odd):nth-child(-n+9)').length is 4
PASS document.querySelectorAll('#test-root :nth-child(n+2):nth-child(odd):nth-child(-n+9)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(n+2):nth-child(odd):nth-child(-n+9)')[1].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(n+2):nth-child(odd):nth-child(-n+9)')[2].id is "testcase7"
PASS document.querySelectorAll('#test-root :nth-child(n+2):nth-child(odd):nth-child(-n+9)')[3].id is "testcase9"
PASS styledElements is ["testcase3","testcase5","testcase7","testcase9"]

Testing ":nth-child(3n+1):nth-child(even)"
PASS document.querySelectorAll('#test-root :nth-child(3n+1):nth-child(even)').length is 2
PASS document.querySelectorAll('#test-root :nth-child(3n+1):nth-child(even)')[0].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(3n+1):nth-child(even)')[1].id is "testcase10"
PASS styledElements is ["testcase4","testcase10"]

Testing ":nth-child(even):nth-child(3n+1)"
PASS document.querySelectorAll('#test-root :nth-child(even):nth-child(3n+1)').length is 2
PASS document.querySelectorAll('#test-root :nth-child(even):nth-child(3n+1)')[0].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(even):nth-child(3n+1)')[1].id is "testcase10"
PASS styledElements is ["testcase4","testcase10"]

Testing ":nth-child(3)"
PASS document.querySelectorAll('#test-root :nth-child(3)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(3)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing ":nth-child(4)"
PASS document.querySelectorAll('#test-root :nth-child(4)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(4)')[0].id is "testcase4"
PASS styledElements is ["testcase4"]

Testing ":nth-child(3):nth-child(4)"
PASS document.querySelectorAll('#test-root :nth-child(3):nth-child(4)').length is 0
PASS styledElements is []

Testing ":nth-child(4):nth-child(3)"
PASS document.querySelectorAll('#test-root :nth-child(4):nth-child(3)').length is 0
PASS styledElements is []

Testing ":nth-child(n+3)"
PASS document.querySelectorAll('#test-root :nth-child(n+3)').length is 9
PASS document.querySelectorAll('#test-root :nth-child(n+3)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(n+3)')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(n+3)')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(n+3)')[3].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(n+3)')[4].id is "testcase7"
PASS document.querySelectorAll('#test-root :nth-child(n+3)')[5].id is "testcase8"
PASS document.querySelectorAll('#test-root :nth-child(n+3)')[6].id is "testcase9"
PASS document.querySelectorAll('#test-root :nth-child(n+3)')[7].id is "testcase10"
PASS document.querySelectorAll('#test-root :nth-child(n+3)')[8].id is "testcase11"
PASS styledElements is ["testcase3","testcase4","testcase5","testcase6","testcase7","testcase8","testcase9","testcase10","testcase11"]

Testing ":nth-child(2n+2)"
PASS document.querySelectorAll('#test-root :nth-child(2n+2)').length is 5
PASS document.querySelectorAll('#test-root :nth-child(2n+2)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(2n+2)')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(2n+2)')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(2n+2)')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :nth-child(2n+2)')[4].id is "testcase10"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10"]

Testing ":nth-child(n+3):nth-child(2n+2)"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(2n+2)').length is 4
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(2n+2)')[0].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(2n+2)')[1].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(2n+2)')[2].id is "testcase8"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(2n+2)')[3].id is "testcase10"
PASS styledElements is ["testcase4","testcase6","testcase8","testcase10"]

Testing ":nth-child(2n+2):nth-child(n+3)"
PASS document.querySelectorAll('#test-root :nth-child(2n+2):nth-child(n+3)').length is 4
PASS document.querySelectorAll('#test-root :nth-child(2n+2):nth-child(n+3)')[0].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(2n+2):nth-child(n+3)')[1].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(2n+2):nth-child(n+3)')[2].id is "testcase8"
PASS document.querySelectorAll('#test-root :nth-child(2n+2):nth-child(n+3)')[3].id is "testcase10"
PASS styledElements is ["testcase4","testcase6","testcase8","testcase10"]

Testing ":nth-child(-n+4)"
PASS document.querySelectorAll('#test-root :nth-child(-n+4)').length is 4
PASS document.querySelectorAll('#test-root :nth-child(-n+4)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(-n+4)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(-n+4)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(-n+4)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing ":nth-child(-n+5)"
PASS document.querySelectorAll('#test-root :nth-child(-n+5)').length is 5
PASS document.querySelectorAll('#test-root :nth-child(-n+5)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(-n+5)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(-n+5)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(-n+5)')[3].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(-n+5)')[4].id is "testcase5"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4","testcase5"]

Testing ":nth-child(-n+4):nth-child(-n+5)"
PASS document.querySelectorAll('#test-root :nth-child(-n+4):nth-child(-n+5)').length is 4
PASS document.querySelectorAll('#test-root :nth-child(-n+4):nth-child(-n+5)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(-n+4):nth-child(-n+5)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(-n+4):nth-child(-n+5)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(-n+4):nth-child(-n+5)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing ":nth-child(-n+5):nth-child(-n+4)"
PASS document.querySelectorAll('#test-root :nth-child(-n+5):nth-child(-n+4)').length is 4
PASS document.querySelectorAll('#test-root :nth-child(-n+5):nth-child(-n+4)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(-n+5):nth-child(-n+4)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(-n+5):nth-child(-n+4)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(-n+5):nth-child(-n+4)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing ":nth-child(n+3):nth-child(-n+6)"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(-n+6)').length is 4
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(-n+6)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(-n+6)')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(-n+6)')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(-n+6)')[3].id is "testcase6"
PASS styledElements is ["testcase3","testcase4","testcase5","testcase6"]

Testing ":nth-child(-n+6):nth-child(n+3)"
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(n+3)').length is 4
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(n+3)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(n+3)')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(n+3)')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(n+3)')[3].id is "testcase6"
PASS styledElements is ["testcase3","testcase4","testcase5","testcase6"]

Testing ":nth-child(n+1):nth-child(-n+3)"
PASS document.querySelectorAll('#test-root :nth-child(n+1):nth-child(-n+3)').length is 3
PASS document.querySelectorAll('#test-root :nth-child(n+1):nth-child(-n+3)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(n+1):nth-child(-n+3)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(n+1):nth-child(-n+3)')[2].id is "testcase3"
PASS styledElements is ["testcase1","testcase2","testcase3"]

Testing ":nth-child(-n+3):nth-child(n+1)"
PASS document.querySelectorAll('#test-root :nth-child(-n+3):nth-child(n+1)').length is 3
PASS document.querySelectorAll('#test-root :nth-child(-n+3):nth-child(n+1)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(-n+3):nth-child(n+1)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(-n+3):nth-child(n+1)')[2].id is "testcase3"
PASS styledElements is ["testcase1","testcase2","testcase3"]

Testing ":nth-child(n+3):nth-child(-n+6):nth-child(n+1):nth-child(-n+3)"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(-n+6):nth-child(n+1):nth-child(-n+3)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(-n+6):nth-child(n+1):nth-child(-n+3)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing ":nth-child(-n+6):nth-child(n+3):nth-child(-n+3):nth-child(n+1)"
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(n+3):nth-child(-n+3):nth-child(n+1)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(n+3):nth-child(-n+3):nth-child(n+1)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing ":nth-child(n+3):nth-child(odd):nth-child(-n+6)"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(odd):nth-child(-n+6)').length is 2
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(odd):nth-child(-n+6)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(odd):nth-child(-n+6)')[1].id is "testcase5"
PASS styledElements is ["testcase3","testcase5"]

Testing ":nth-child(-n+6):nth-child(odd):nth-child(n+3)"
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(odd):nth-child(n+3)').length is 2
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(odd):nth-child(n+3)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(odd):nth-child(n+3)')[1].id is "testcase5"
PASS styledElements is ["testcase3","testcase5"]

Testing ":nth-child(odd):nth-child(n+3):nth-child(-n+6)"
PASS document.querySelectorAll('#test-root :nth-child(odd):nth-child(n+3):nth-child(-n+6)').length is 2
PASS document.querySelectorAll('#test-root :nth-child(odd):nth-child(n+3):nth-child(-n+6)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(odd):nth-child(n+3):nth-child(-n+6)')[1].id is "testcase5"
PASS styledElements is ["testcase3","testcase5"]

Testing ":nth-child(odd):nth-child(-n+6):nth-child(n+3)"
PASS document.querySelectorAll('#test-root :nth-child(odd):nth-child(-n+6):nth-child(n+3)').length is 2
PASS document.querySelectorAll('#test-root :nth-child(odd):nth-child(-n+6):nth-child(n+3)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(odd):nth-child(-n+6):nth-child(n+3)')[1].id is "testcase5"
PASS styledElements is ["testcase3","testcase5"]

Testing ":nth-child(n+3):nth-child(-n+6):nth-child(odd)"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(-n+6):nth-child(odd)').length is 2
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(-n+6):nth-child(odd)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(-n+6):nth-child(odd)')[1].id is "testcase5"
PASS styledElements is ["testcase3","testcase5"]

Testing ":nth-child(-n+6):nth-child(n+3):nth-child(odd)"
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(n+3):nth-child(odd)').length is 2
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(n+3):nth-child(odd)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(-n+6):nth-child(n+3):nth-child(odd)')[1].id is "testcase5"
PASS styledElements is ["testcase3","testcase5"]

Testing ":nth-child(n+1):nth-child(even):nth-child(-n+3)"
PASS document.querySelectorAll('#test-root :nth-child(n+1):nth-child(even):nth-child(-n+3)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(n+1):nth-child(even):nth-child(-n+3)')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":nth-child(-n+3):nth-child(even):nth-child(n+1)"
PASS document.querySelectorAll('#test-root :nth-child(-n+3):nth-child(even):nth-child(n+1)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(-n+3):nth-child(even):nth-child(n+1)')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":nth-child(even):nth-child(n+1):nth-child(-n+3)"
PASS document.querySelectorAll('#test-root :nth-child(even):nth-child(n+1):nth-child(-n+3)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(even):nth-child(n+1):nth-child(-n+3)')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":nth-child(even):nth-child(-n+3):nth-child(n+1)"
PASS document.querySelectorAll('#test-root :nth-child(even):nth-child(-n+3):nth-child(n+1)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(even):nth-child(-n+3):nth-child(n+1)')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":nth-child(-n+3):nth-child(n+1):nth-child(even)"
PASS document.querySelectorAll('#test-root :nth-child(-n+3):nth-child(n+1):nth-child(even)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(-n+3):nth-child(n+1):nth-child(even)')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":nth-child(n+1):nth-child(-n+3):nth-child(even)"
PASS document.querySelectorAll('#test-root :nth-child(n+1):nth-child(-n+3):nth-child(even)').length is 1
PASS document.querySelectorAll('#test-root :nth-child(n+1):nth-child(-n+3):nth-child(even)')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":nth-child(n+3):nth-child(odd):nth-child(-n+6):nth-child(n+1):nth-child(-n+3):nth-child(even)"
PASS document.querySelectorAll('#test-root :nth-child(n+3):nth-child(odd):nth-child(-n+6):nth-child(n+1):nth-child(-n+3):nth-child(even)').length is 0
PASS styledElements is []

PASS successfullyParsed is true

TEST COMPLETE
